+++ server.js
@@ -4,20 +4,21 @@ const cron = require("node-cron");
 const express = require("express");
 const cors = require("cors");
 const morgan = require("morgan");
+const mongoose = require("mongoose");
 const swaggerJsDoc = require("swagger-jsdoc");
 const swaggerUi = require("swagger-ui-express");
 
-const { connect: connectDB } = require("./lib/db");
-const { fetchAndSaveTweets } = require("./lib/fetchTweets");
+const { fetchAndSaveTweets } = require("./fetchTweets");
+const categoryController = require("./controllers/category");
 
 const categoriesRoutes = require("./routes/categories");
 const tweetsRoutes = require("./routes/tweets");
 const metaRoutes = require("./routes/meta");
 const verificationRoutes = require("./routes/verification");
-const feedbackRoutes = require("./routes/feedback");
 
 const app = express();
 
+const DB_URL = process.env.MONGO_URI;
 const PORT = process.env.PORT || 4000;
 
 const swaggerDocs = swaggerJsDoc({
@@ -46,7 +47,6 @@ app.use("/api", tweetsRoutes);
 app.use("/api", categoriesRoutes);
 app.use("/api", metaRoutes);
 app.use("/api", verificationRoutes);
-app.use("/api", feedbackRoutes);
 
 app.use("/", swaggerUi.serve, swaggerUi.setup(swaggerDocs));
 console.log(
@@ -54,25 +54,29 @@ console.log(
   process.env.NODE_ENV == "production" ? "prod" : "staging",
   " Environment"
 );
+mongoose
+  .connect(DB_URL, { dbName: process.env.NODE_ENV == "production" ? "prod" : "staging",useNewUrlParser: true, useUnifiedTopology: true })
+  .then(() => {
+    console.log("✅ Database Connected!");
 
-connectDB().then(() => {
-  console.log("✅ Database Connected!");
+    fetchAndSaveTweets();
 
-  fetchAndSaveTweets();
+    if (process.env.NODE_ENV === "production") {
+      cron.schedule("*/5 * * * *", async () => {
+        console.log("Fetching Tweets...");
+        console.time("fetchTweets");
 
-  if (process.env.NODE_ENV === "production") {
-    cron.schedule("*/1 * * * *", async () => {
-      console.log("Fetching Tweets...");
-      console.time("fetchTweets");
+        await fetchAndSaveTweets();
 
-      await fetchAndSaveTweets();
+        console.timeEnd("fetchTweets");
+        console.log("Done Fetching Tweets!");
+      });
+    }
 
-      console.timeEnd("fetchTweets");
-      console.log("Done Fetching Tweets!");
+    app.listen(PORT, "0.0.0.0", () => {
+      console.log("🚀 Server Ready! at port:", PORT);
     });
-  }
-
-  app.listen(PORT, "0.0.0.0", () => {
-    console.log("🚀 Server Ready! at port:", PORT);
+    process.on("beforeExit", () => {
+      categoryController.flush();
+    });
   });
-});
